SystemUI: Enable three icon switching within QS DND tile
This expands the QS DND tile to use three states instead of two.

Java commits by @nicholaschum

What this does:
- Removes the two states (true/false) clause {? XXX:YYY;} and uses an
switch(mZen) instead for the two cases and the default. If in the
future, we want to add another tile, just add another case.
- Adds two vector drawables originally by Cataclysm dev @atl4ntis,
modified and reformatted - this allows for themers to theme this
resource as well by adding the two new drawables.

Change-Id: I1bee47bc04b2536d2d20cca4d7a6009067893a4a
PS1: Uses if/elsif/else clause
PS2: Update to use switch(mZen) clause
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>