DocumentsUI: use "Files" application name properly
After we have switched to productivity device externally, there's still
an issue for non-english languages that won't be able to have the proper
launcher and files activities with the proper name for the new feature.

Work around this behaviour switching to "app_label" string for both
activities. This makes the usage of the previous work around in vendor
not needed anymore.

Change-Id: I4a3e273fe07e5107257377a2335401f75ede5900
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: spezi77 <spezi77@gmx.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>